# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Phoenix Guardian API
  description: |
    Phoenix Guardian is an AI-powered healthcare security platform that provides
    real-time threat detection, secure transcription, and SOAP note generation
    for medical encounters.
    
    ## Authentication
    
    All API endpoints require authentication using JWT Bearer tokens.
    Tokens are obtained from the `/auth/token` endpoint and must include
    the tenant ID in the claims.
    
    ## Multi-Tenancy
    
    Phoenix Guardian is a multi-tenant platform. Each request must include
    the tenant context via the JWT token or X-Tenant-ID header.
    
    ## Rate Limiting
    
    - Standard: 1000 requests/minute per tenant
    - Transcription: 100 requests/minute per tenant
    - AI Operations: 50 requests/minute per tenant
    
    ## Versioning
    
    The API is versioned via the URL path (e.g., `/api/v1/`).
    
  version: 3.0.0
  contact:
    name: Phoenix Guardian API Support
    email: api-support@phoenix-guardian.health
    url: https://docs.phoenix-guardian.health
  license:
    name: Proprietary
    url: https://phoenix-guardian.health/license
servers:
  - url: https://api.phoenix-guardian.health/api/v1
    description: Production
  - url: https://staging-api.phoenix-guardian.health/api/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local Development
...
security:
  - BearerAuth: []
paths:
...
  /auth/refresh:
    post:
      summary: Refresh access token
      description: Exchange refresh token for new access token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required:
                - refresh_token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      summary: Logout and invalidate tokens
      tags:
        - Authentication
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==========================================================================
  # Encounters
  # ==========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: Bearer
        expires_in:
          type: integer
          description: Token lifetime in seconds
        scope:
          type: string
          
    # Encounters
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object